+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implemented a Badge component.
 - Implemented a Button component.
 - Implemented a Card component.
+- Implemented a Carousel component.
 - Implemented a Checkbox component.
 - Implemented a Progress component.
 - Implemented a Select component.